Add files via upload
Q. 1) Find all Null Values in the dataset. If there is any null value in any column, then fill it with the mean of that column.

Q. 2) Check what are the different types of Make are there in our dataset. And, what is the count (occurrence) of each Make in the data?

Q. 3) Show all the records where Origin is Asia or Europe.

Q. 4) Remove all the records (rows) where the Weight is above 4000.

Q. 5) Increase all the values of the 'MPG_City' column by 3